// src/types.ts
export interface Discipline {
  id: string;
  name: string;
  iconName: string;
  description: string;
  bgClass: string;
}

export interface MissionItem {
  id: string;
  title: string;
  description: string;
  iconName: string;
}

export interface BenefitItem {
  id: string;
  title: string;
  text: string;
  iconName: string;
}

export interface FeatureItem {
  id: string;
  title: string;
  description: string;
  iconName: string;
}

export interface ValueItem {
  id: string;
  title: string;
  description: string;
  iconName: string;
}

export interface MattersItem {
  id: string;
  title: string;
  description: string;
  iconName: string;
  details: string;
}